GetTickCount now depends on the TSC
This is a major change in terms of development, and you may wonder
"why?".

Before this commit, the millisecond timer used to be accurate only
down to like 2 milliseconds or something. Now, it's pretty much
accurate down to the tens of microseconds, which is really helpful
for tracking CPU time spent running code. Eventually I'll write an
actual CPU usage tracker (probably in the SystemMonitor), but that'll
have to wait.

Oh, and also, I added '%q' and '%Q' in printf(), which print a 64-bit
unsigned integer in hexadecimal (%q) and decimal (%Q) form.