Wm: Optimize obscured and foremost checks further
Now the depth buffer refresh is the one that decides what windows
are foremost / obscured.

A foremost window is a window which has its entire drawing area visible
on the screen. This allows us to optimize out any occlusion checks we
would do otherwise.  In fact, this is what we've been doing for a long
time, but now, the foremost/obscured results are only recalculated on
a depth buffer refresh instead of every single time a window requests
a refresh.  This should make RenderWindow way faster to refresh foremost
windows.

An obscured window is a window which has no parts of its drawing area
visible. This is usually because one or more windows are on top of it.
There's no point in trying to look for an unoccluded pixel and drawing
it when there ARE no pixels to draw. It's common sense to not even try.

This commit also does a lot of other changes:
- changed SystemMonitor icon
- modified some of SystemMonitor's layout code
- the terminal now spawns using CW_AUTOPOSITION