Vid: Adding the first version of Dirty rect tracking
Ok, so, first of all, how to toggle dirty rect tracking
in your VBEdata? All you have to do is set the first byte
(before, it was `m_available`, now it's actually `m_version`)
to 2 (VBEDATA_VERSION_2) to enable this feature, and allocate
some space after m_clipRect for the DsjRectSet.  The DirtyRectLogger
will automatically get called, and log your dirty rect trackings
there.

If the dirty rect readings are SPARSE enough (i.e. they can't be
be merged automatically by the dirty rect handler, and they overflow
the currently 64-limit rectangle container), the handler will say
"eh, screw it, just mark _everything_ as dirty, it will hardly
make a difference.

I've used the `cli` and `sti` instructions to prevent data races
coming from this, however, I think there should be a better method.
Perhaps use a `SafeLock` instead?

In order to test this out, press F11, which will now corrupt your
screen. This is so I can test the feature out, if it works or not.
This does NOT corrupt any of the windows themselves (as an unfocus
will restore all the detail previously corrupted), but it allows
the user to see how the dirty rects work.

I've discovered one major flaw, which I'll work on SOON (this is a
TODO), which is that a big rectangle that covers a large part of the
screen will just get merged together to be the screen itself. Kinda
defeats the purpose eh?  I'll probably fix it later. (tm)

It's actually still kind of rusty at times, but I hope I'll get this
stable enough soon.