Mm: Add a more efficient MmReAllocate
While this MmReAllocate - as a last resort - MmAllocates, memcpies
and MmFrees the old_ptr and returns a whole new address, it usually
isn't the case. This function tries to resize the memory block as-is
to avoid copying memory uselessly.

Potentially I could also optimize the more inefficient method by
instead mapping the pages themselves somewhere else instead of allocating
a whole new space.