* Switch mainline NanoShell to the new memory manager.
Ok, we've known for a while now that NanoShell needed a new memory
manager. So much so, that on August 11th, 2022, I've started a new
repository to experiment with creating a new memory manager from
(almost) scratch. Now that I consider it done, this commit will
add it to NanoShell, and retire the original one. You served us well,
old friend.

Sadly this does come with a few shortcomings which I will address in
the coming days/weeks:
- MmAllocate, MmFree only works within interrupt-enabled functions.
  This is actually OK since the old ones also worked only there,
  using them in interrupt-disabled functions would cause data races
  and all kinds of bad stuff.
- MmAllocate and MmFree only work within the *kernel heap*. I plan
  on ditching the legacy APIs anyway in favor of better control on
  the user side (I'll probably port a malloc to the NanoShell CRT)
- Window manager performance is reduced (namely when creating/
  resizing windows), because we have to fault a bunch to bring the
  pages to fruition. This will be fixed eventually.
- There may still be some assertions and stuff left for me to fix
  (interrupt-disabled places calling into interrupt-enabled-requiring
  functions etc)

This has a few upsides too though, so I see this as a positive move:
- The user has better control over the malloc implementation that's
  being used (be it dlmalloc, mimalloc or whatever have you)
- This fixes a bunch of data races caused by uncareful memory manipu-
  lation (For instance, KeStartTaskEx has a random chance on crashing
  upon scheduling the task itself when iret-ing)
- Makes it easier overall to do stuff such as physical memory mapping
  (it's part of the kernel heap now) and Copy-on-write (probably won't
  be used much though unless we add fork() in the future)

I hope this serves us well!
- iProgramInCpp