* Fix memory leak with ELF executive when program crashes
The leak was caused by our careless attempt at mitigating a bug.
In the 'before' times, NanoShell had a kernel heap and a user heap,
and things could allocate to both. Now, however, we have just the
kernel heap, the user heap being MemoryMap/MemoryUnMap-only (so,
managed by the individual program itself)... (although that may not
have been the best call :^) to be honest, since it inflated the file
size quite a lot).

Don't make a local kernel-heap copy of the file data, instead keep it
as-is, and free it in the OnDeath function.