* Resource: Add LaunchFileOrResource.
LaunchFileOrResource firstly calls LaunchResource. If it fails,
with an error code other than RESOURCE_LAUNCH_NO_PROTOCOL, it's done.

If there's no actual protocol, we have a file association table
(static for now), which we can look through and check. We use said
table to determine the protocol to open the file (the way we should
open it).