* Allow changing of the hue/sat/val properties.
This also changes the scale from 0-100 to 0-256, to mitigate
potential imprecisions. Assuming a precision of 2^14, (2^14)*0.1
== 1638.4 (10%). Since it has to be rounded down, it's rounded to
0.099... This is bad since it makes it show 9% instead of 10%.
This problem is even worse with a 2^12 precision.

There's no such imprecision in the hue scale, since it uses 0-360
instead of 0 to 1 scale. So you just work with multiples of 2^PREC,
instead of parts of 2^PREC.

Ideally, I'd just change the HSVToRGB function to take a scale of
0-100 instead of 0 to 1, but this is fine too. Unfortunately, it takes
some conversion from a usual HSV scale to get the right color.

Also, I may want to add an RGBToHSV function as well, since I'd eventually
like to also be able to change the RGB fields. They don't do anything
right now.