* Revamp crash handler to queue up crashes.
This gets around a really small edge cases wherein two crashes
crash at roughly the same time, and the crash reporter doesn't
process the first crash fast enough, so the second one is just
thrown up on the screen and stops the whole system.

Also this desperately needed a revamp. :)