* Add MmAllocateID and MmFreeID.
This ends the time where "mm/memoryi.h" was included directly
for operations where the interrupts are disabled. Since there
are a bunch of cases where the slab allocator intervenes only
when allocating, and not when freeing, I've decided to make the
slab allocator not use a lock, and instead, depend on interrupts
being disabled.